+++ benchopt/tests/test_benchmarks.py
@@ -54,9 +54,10 @@ def test_dataset_get_data(benchmark, dataset_class):
     config = get_configs(dataset_class)[0]
     dataset = dataset_class.get_instance(**config['dataset'])
 
-    # ensure classes calling `get_seed` work properly
+    # ensure classes calling `get_seed` work properly. Use placeholder
+    # objective/solver names to mimic known dimension at runtime.
     RunContext().set_run_context(
-        None, dataset, None, repetition=0, base_seed=0
+        "dummy", dataset, "dummy", repetition=0, base_seed=0
     )
 
     data = dataset._get_data()
@@ -80,9 +81,10 @@ def test_benchmark_objective(benchmark, test_dataset_name):
     dataset = dataset_class.get_instance(**config['dataset'])
     objective = objective_class.get_instance(**config['objective'])
 
-    # ensure classes calling `get_seed` work properly
+    # ensure classes calling `get_seed` work properly. Use a placeholder
+    # solver name so all methods can use any seed dimension.
     RunContext().set_run_context(
-        objective, dataset, None, repetition=0, base_seed=0
+        objective, dataset, "dummy", repetition=0, base_seed=0
     )
 
     # get one value for the objective, with the test_dataset
@@ -216,10 +218,11 @@ def test_solver_stopping_criterion(benchmark, solver_class, test_dataset_name):
             "'objective_'."
         )
 
-        # ensure classes calling `get_seed` work properly
+        # ensure classes calling `get_seed` work properly.
         dataset = dataset_class.get_instance(**config['dataset'])
         RunContext().set_run_context(
-            objective, dataset, None, repetition=0, base_seed=benchmark.seed
+            objective, dataset, solver_class.name, repetition=0,
+            base_seed=benchmark.seed
         )
 
         objective._set_dataset(dataset)